FIX: Lightning and Vault Icons
Lightning and Vault icons seem to be cut off from the top. I have noticed that the height for the Bitcoin icon is 381 pixels, and it is the only one that is displayed correctly. Since I don’t know how to fix this with code (but you probably do), here are all of the icons with the same height. The benefit now is that they all look the same size.