Additional accessibility fixes
I used the "eslint-plugin-react-native-a11y" plug-in to identify more problems and went through all the commendations for buttons. I ignored input fields (about 25 issues found), since I am not 100% sure about the best way to fix them and it might be better to tackle that in a separate PR and keep this one focused on buttons.